requirements.docker.txt
fastapi==0.123.8
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
pydantic==2.12.5
openai==2.14.0
psycopg[binary,pool]==3.3.2
redis==7.1.0
twilio==9.6.4
python-multipart==0.0.21
numpy==2.4.4
tiktoken==0.12.0
tqdm==4.67.1
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
python-pptx==1.0.2
httpx==0.28.1
playwright==1.58.0
